Fixes for Spring upgrade
