+++ src/main/java/ch/wisv/areafiftylan/Application.java
@@ -19,15 +19,13 @@ package ch.wisv.areafiftylan;
 
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @Configuration
-@ComponentScan
-@EnableAutoConfiguration
 @EnableScheduling
+@SpringBootApplication
 public class Application {
 
     public static void main(String[] args) {
