+++ src/main/resources/config/application.yml
@@ -1,5 +1,7 @@
 ## SPRING SETTINGS
 spring:
+    main:
+        allow-bean-definition-overriding: true
     thymeleaf:
         check-template-location: true
         prefix: classpath:/templates/
